requirements.txt
click>=7.0
orjson>=3.4
pandas>=1.1
ply>=3.11
python-dotenv>=0.10
requests>=2.21
rich>=9.2
smart-open>=3.0
sqlalchemy>=1.3
